refactor: change directory name from `utils` to `lib`
